Camelize strategy name in shared examples
Following up on #681, we also need to camelize strategy names
in the RSpec shared examples for cleaning strategies.

For developer convenience, the methods for string manipulation
were moved from the Cleaner class to a dedicated StringHelper module.
