+++ lib/database_cleaner/spec/shared_examples.rb
@@ -11,7 +11,7 @@ end
 RSpec.shared_examples_for "a database_cleaner adapter" do
   describe 'all strategies should adhere to a database_cleaner strategy interface' do
     DatabaseCleaner::Cleaner.available_strategies(described_class).each do |strategy|
-      subject { described_class.const_get(strategy.to_s.capitalize).new }
+      subject { described_class.const_get(DatabaseCleaner::StringHelper.camelize(strategy.to_s)).new }
 
       it_behaves_like 'a database_cleaner strategy'
     end
